Removed spark dependency - not needed Moved test to src folder for sbt correctness Fixed first two tests - sum / onlyEvenNumbers
