+++ build.sbt
@@ -5,5 +5,5 @@ version := "0.1"
 scalaVersion := "2.12.0"
 
 libraryDependencies ++= Seq(
-  "org.apache.spark" %% "spark-sql" % "2.4.3",
-  "org.scalatest" %% "scalatest" % "3.0.4")
+  "org.scalatest" %% "scalatest" % "3.0.4"
+)
